chore: shared types refactor
