+++ packages/shared-types/src/pdf-stream.ts
@@ -1,334 +1,3 @@
-import { z } from "zod";
-
-const JsonRecordSchema = z.record(z.string(), z.unknown());
-
-export const PdfParseJobDataSchema = z.strictObject({
-  jobId: z.string().trim().min(1),
-  fileKey: z.string().trim().min(1),
-  adapterKey: z.string().trim().min(1),
-});
-
-export type PdfParseJobData = z.infer<typeof PdfParseJobDataSchema>;
-
-export const ParseAnnotationSchema = z.strictObject({
-  code: z.string(),
-  message: z.string(),
-  details: JsonRecordSchema,
-});
-
-export type ParseAnnotation = z.infer<typeof ParseAnnotationSchema>;
-
-export const ParsedModuleCandidateSchema = z.strictObject({
-  code: z.string(),
-  name: z.string().nullable(),
-  metadata: JsonRecordSchema,
-  warnings: z.array(ParseAnnotationSchema),
-});
-
-export type ParsedModuleCandidate = z.infer<typeof ParsedModuleCandidateSchema>;
-
-export const ActivityTypeSchema = z.enum([
-  "lecture",
-  "tutorial",
-  "prac",
-  "test",
-  "exam",
-]);
-
-export type ActivityType = z.infer<typeof ActivityTypeSchema>;
-
-export const DayOfWeekSchema = z.enum([
-  "monday",
-  "tuesday",
-  "wednesday",
-  "thursday",
-  "friday",
-  "saturday",
-  "sunday",
-]);
-
-export type DayOfWeek = z.infer<typeof DayOfWeekSchema>;
-
-const ParsedDayAliasSchema = z.enum([
-  "mon",
-  "monday",
-  "tue",
-  "tues",
-  "tuesday",
-  "wed",
-  "wednesday",
-  "thu",
-  "thur",
-  "thurs",
-  "thursday",
-  "fri",
-  "friday",
-  "sat",
-  "saturday",
-  "sun",
-  "sunday",
-]);
-
-const parsedDayAliases = {
-  mon: "monday",
-  monday: "monday",
-  tue: "tuesday",
-  tues: "tuesday",
-  tuesday: "tuesday",
-  wed: "wednesday",
-  wednesday: "wednesday",
-  thu: "thursday",
-  thur: "thursday",
-  thurs: "thursday",
-  thursday: "thursday",
-  fri: "friday",
-  friday: "friday",
-  sat: "saturday",
-  saturday: "saturday",
-  sun: "sunday",
-  sunday: "sunday",
-} as const satisfies Record<z.infer<typeof ParsedDayAliasSchema>, DayOfWeek>;
-
-export const ParsedDayOfWeekSchema = z
-  .string()
-  .trim()
-  .toLowerCase()
-  .pipe(ParsedDayAliasSchema)
-  .transform((day) => parsedDayAliases[day]);
-
-const TimeOfDaySchema = z.string().regex(/^(?:[01]\d|2[0-3]):[0-5]\d$/, {
-  message: "Expected a time in HH:mm format.",
-});
-
-export const IsoDateSchema = z
-  .string()
-  .regex(/^\d{4}-\d{2}-\d{2}$/, {
-    message: "Expected an ISO calendar date (YYYY-MM-DD).",
-  })
-  .refine(
-    (value) => {
-      const date = new Date(`${value}T00:00:00.000Z`);
-      return (
-        !Number.isNaN(date.getTime()) &&
-        date.toISOString().slice(0, 10) === value
-      );
-    },
-    { message: "Expected a real ISO calendar date." },
-  );
-
-const ParsedEventCandidateBaseShape = {
-  moduleCode: z.string(),
-  activityType: ActivityTypeSchema,
-  activityCode: z.string(),
-  title: z.string(),
-  startTime: z.string(),
-  endTime: z.string(),
-  venues: z.array(z.string()),
-  metadata: JsonRecordSchema,
-  warnings: z.array(ParseAnnotationSchema),
-};
-
-export const ParsedEventCandidateSchema = z.discriminatedUnion("isRecurring", [
-  z.strictObject({
-    ...ParsedEventCandidateBaseShape,
-    day: ParsedDayOfWeekSchema,
-    date: z.null(),
-    isRecurring: z.literal(true),
-  }),
-  z.strictObject({
-    ...ParsedEventCandidateBaseShape,
-    day: z.null(),
-    date: IsoDateSchema,
-    isRecurring: z.literal(false),
-  }),
-]);
-
-export type ParsedEventCandidate = z.infer<typeof ParsedEventCandidateSchema>;
-
-export const PdfParserResultSchema = z.strictObject({
-  modules: z.array(ParsedModuleCandidateSchema),
-  events: z.array(ParsedEventCandidateSchema),
-  warnings: z.array(ParseAnnotationSchema),
-});
-
-export type PdfParserResult = z.infer<typeof PdfParserResultSchema>;
-
-export const WorkerCallbackErrorSchema = z.strictObject({
-  code: z.string(),
-  message: z.string(),
-  details: JsonRecordSchema.optional(),
-});
-
-export type WorkerCallbackError = z.infer<typeof WorkerCallbackErrorSchema>;
-
-export const PdfParserCallbackPayloadSchema = z.discriminatedUnion("status", [
-  z.strictObject({
-    status: z.literal("completed"),
-    result: PdfParserResultSchema,
-  }),
-  z.strictObject({
-    status: z.literal("failed"),
-    error: WorkerCallbackErrorSchema,
-  }),
-]);
-
-export type PdfParserCallbackPayload = z.infer<
-  typeof PdfParserCallbackPayloadSchema
->;
-
-export const SolverEngineSchema = z.enum(["auto", "cp-sat", "ga"]);
-
-export type SolverEngine = z.infer<typeof SolverEngineSchema>;
-
-export const SolverHeuristicPreferenceSchema = z.strictObject({
-  key: z.string().trim().min(1),
-  weight: z.number().finite().optional(),
-  parameters: JsonRecordSchema.optional(),
-});
-
-export type SolverHeuristicPreference = z.infer<
-  typeof SolverHeuristicPreferenceSchema
->;
-
-export const SolverPreferencesSchema = z.strictObject({
-  heuristics: z.array(SolverHeuristicPreferenceSchema).default([]),
-});
-
-export type SolverPreferences = z.infer<typeof SolverPreferencesSchema>;
-
-export const TimetableSolveJobDataSchema = z.strictObject({
-  jobId: z.string().trim().min(1),
-  attemptToken: z.uuid(),
-  solveMode: z.enum(["feasibility", "optimization"]),
-  engine: SolverEngineSchema.default("auto"),
-});
-
-export type TimetableSolveJobData = z.infer<typeof TimetableSolveJobDataSchema>;
-
-/**
- * The backend owns the problem model. Heuristics are soft ranking inputs and
- * never change whether a solution is valid.
- */
-export const SchedulingEventSchema = z
-  .strictObject({
-    eventId: z.string().trim().min(1),
-    moduleCode: z.string().trim().min(1),
-    activityType: ActivityTypeSchema,
-    activityCode: z.string().trim().min(1),
-    requiredSelections: z.number().int().positive().default(1),
-    date: IsoDateSchema.optional(),
-    dayOfWeek: DayOfWeekSchema.optional(),
-    startTime: TimeOfDaySchema,
-    endTime: TimeOfDaySchema,
-    venues: z
-      .array(
-        z.strictObject({
-          id: z.string().trim().min(1),
-          name: z.string().trim().min(1),
-        }),
-      )
-      .default([]),
-  })
-  .superRefine((event, context) => {
-    if (Boolean(event.date) === Boolean(event.dayOfWeek)) {
-      context.addIssue({
-        code: "custom",
-        message: "Scheduling events require exactly one of date or dayOfWeek.",
-      });
-    }
-
-    if (event.startTime >= event.endTime) {
-      context.addIssue({
-        code: "custom",
-        path: ["endTime"],
-        message: "endTime must be later than startTime.",
-      });
-    }
-  });
-
-export type SchedulingEvent = z.infer<typeof SchedulingEventSchema>;
-
-export const SchedulingProblemSchema = z.strictObject({
-  events: z.array(SchedulingEventSchema),
-});
-
-export type SchedulingProblem = z.infer<typeof SchedulingProblemSchema>;
-
-export const TimetableSolutionSchema = z.strictObject({
-  selectedEventIds: z.array(z.string().trim().min(1)),
-});
-
-export type TimetableSolution = z.infer<typeof TimetableSolutionSchema>;
-
-export const SolverInputSchema = z.strictObject({
-  schedulingProblem: SchedulingProblemSchema,
-  preferences: SolverPreferencesSchema.default({ heuristics: [] }),
-});
-
-export type SolverInput = z.infer<typeof SolverInputSchema>;
-
-export const SolverHeuristicScoreSchema = z.strictObject({
-  key: z.string().trim().min(1),
-  score: z.number().finite(),
-  details: JsonRecordSchema.optional(),
-});
-
-export type SolverHeuristicScore = z.infer<typeof SolverHeuristicScoreSchema>;
-
-export const SolverConflictSchema = z.strictObject({
-  eventIds: z.tuple([z.string().trim().min(1), z.string().trim().min(1)]),
-});
-
-export const SolverResultMetadataSchema = z
-  .object({
-    conflictCount: z.number().int().nonnegative(),
-    conflicts: z.array(SolverConflictSchema),
-    solveMode: z.enum(["feasibility", "optimization"]),
-  })
-  .catchall(z.unknown());
-
-export const SolverResultSchema = z
-  .strictObject({
-    engine: z.enum(["cp-sat", "ga"]),
-    outcome: z.enum(["conflict-free", "best-effort"]),
-    timetableSolution: TimetableSolutionSchema,
-    heuristicScores: z.array(SolverHeuristicScoreSchema).default([]),
-    metadata: SolverResultMetadataSchema,
-  })
-  .superRefine((result, context) => {
-    if (result.metadata.conflictCount !== result.metadata.conflicts.length) {
-      context.addIssue({
-        code: "custom",
-        path: ["metadata", "conflictCount"],
-        message: "conflictCount must equal the number of conflicts.",
-      });
-    }
-    const expectedOutcome =
-      result.metadata.conflictCount === 0 ? "conflict-free" : "best-effort";
-    if (result.outcome !== expectedOutcome) {
-      context.addIssue({
-        code: "custom",
-        path: ["outcome"],
-        message: `outcome must be ${expectedOutcome} for this conflict count.`,
-      });
-    }
-  });
-
-export type SolverResult = z.infer<typeof SolverResultSchema>;
-
-export const SolverCallbackPayloadSchema = z.discriminatedUnion("status", [
-  z.strictObject({
-    status: z.literal("completed"),
-    result: SolverResultSchema,
-  }),
-  z.strictObject({
-    status: z.literal("failed"),
-    error: WorkerCallbackErrorSchema,
-  }),
-]);
-
-export type SolverCallbackPayload = z.infer<typeof SolverCallbackPayloadSchema>;
-
 export const PDF_STREAM_FINGERPRINT_ALGORITHM_VERSION =
   "pdf-stream-payload-sha256-v1" as const;
